+++ libs/private/functions/payments/core/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/receive-payments.handler';
+export * from './lib/handlers/receive-payments.handler';
 export * from './lib/commands/receive-payment.command';
-export * from './lib/request-payment.handler';
-export * from './lib/subscription-checker.handler';
-export * from './lib/create-customer-subscription.handler'
-
+export * from './lib/handlers/request-payment.handler';
+export * from './lib/handlers/subscription-checker.handler';
+export * from './lib/handlers/create-customer-subscription.handler'
+export * from './lib/handlers/create-mollie-customer.handler'
